debug 季報計算會出現無限大或除以零的狀況

+++ handler/addSeason.ashx
@@ -457,8 +457,13 @@ public class addSeason : IHttpHandler,IRequiresSessionState {
         string tmpMoney = (Money == "") ? "0" : Money;
         double m = double.Parse(tmpMoney);
         double rm = double.Parse(tmpReal);
-        ratio = Math.Round(((rm / m) * 100), 0).ToString();
-        return ratio;
+        if (m == 0)
+            return "0";
+        else
+        {
+            ratio = Math.Round(((rm / m) * 100), 0).ToString();
+            return ratio;
+        }
     }
